tests: Move the @params tests into test_ocs_agent.py
Also translate them to py.test framework.
